# docs/swagger.yaml
basePath: /api
definitions:
...
  menu.SDbMenu:
    properties:
      access:
        type: integer
      icon:
        type: string
      keterangan:
        type: string
      kodemenu:
        type: string
      l0:
        type: integer
      ol:
        type: integer
      platformMask:
        type: integer
      routename:
        type: string
      tipeTrans:
        type: string
    type: object
...
  user.SCreateUserPayload:
    properties:
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - employee
        - karyawan
        type: string
      username:
        minLength: 3
        type: string
    required:
    - full_name
    - password
    - role
    - username
    type: object
  user.SUpdateUserPayload:
    properties:
      full_name:
        type: string
      password:
        description: 'optional: leave blank to keep unchanged'
        type: string
      role:
        enum:
        - admin
        - employee
        - karyawan
        type: string
      status:
        description: 'optional: "0" = inactive, "1" = active'
        type: string
    required:
    - full_name
    - role
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for DAPEN management of employee retirement funds
  title: DAPEN System API
  version: "1.0"
paths:
...
  /admin/menu:
    get:
      description: Fetch a paginated list of all menus
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Menus
      tags:
      - Menu
    post:
      consumes:
      - application/json
      description: Add a new menu item to the system
      parameters:
      - description: Menu Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/menu.SDbMenu'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create New Menu
      tags:
      - Menu
  /admin/menu/{kode}:
    delete:
      description: Remove a menu item from the system
      parameters:
      - description: Menu Code
        in: path
        name: kode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Menu
      tags:
      - Menu
    get:
      description: Fetch a single menu by its code
      parameters:
      - description: Menu Code
        in: path
        name: kode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Menu By ID
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: Modify an existing menu item
      parameters:
      - description: Menu Code
        in: path
        name: kode
        required: true
        type: string
      - description: Updated Menu Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/menu.SDbMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu
  /admin/menu/parents:
    get:
      description: Fetch parent menus based on specific hierarchy level
      parameters:
      - description: Menu Level
        in: query
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Parent Menus
      tags:
      - Menu
...
  /api/admin/users:
    get:
      description: Fetch a paginated list of legacy users from the DBFLPASS table
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Hashing password and storing a new user record inside the DBFLPASS
        table
      parameters:
      - description: Create User Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.SCreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create User
      tags:
      - User
  /api/admin/users/{id}:
    delete:
      description: Delete a user from the DBFLPASS table and prevent self-deletion
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update legacy fields for a user inside the DBFLPASS table
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update User Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.SUpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update User
      tags:
      - User
...
swagger: "2.0"